refactor: main에서 컨트롤러로 로직 이동
