+++ src/main/kotlin/racingcar/GameStarter.kt
@@ -1,22 +1,8 @@
 package racingcar
 
-import racingcar.model.CarInitializer
-import racingcar.model.RacingCarGame
-import racingcar.strategy.MovingStrategy
-import racingcar.strategy.RandomNumberMovingStrategy
-import racingcar.view.InputView
-import racingcar.view.ResultView
+import racingcar.controller.RacingCarController
 
 fun main() {
-    val strategy: MovingStrategy = RandomNumberMovingStrategy()
-    val numbersOfCar = InputView.getNumbersOfCars()
-    val numbersOfTry = InputView.getNumbersOfTry()
-    val cars = CarInitializer.createCars(numbersOfCar)
-    val game = RacingCarGame(cars, strategy)
-    ResultView.printInitial()
-    for (i in 1..numbersOfTry) {
-        game.run()
-        ResultView.printResult(game.cars)
-        ResultView.printEmptyLine()
-    }
+    val racingCarController = RacingCarController()
+    racingCarController.startGame()
 }
